[BUGFIX] Exclude Squiz.PHP.Eval-Sniff for UnitTests
In unit tests of TYPO3v4 and FLOW3, the eval() function is used
for to build accessible mocks.
Conclusion: Disable Squiz.PHP.Eval" for unit tests, but all
other sniffs are enabled, because the quality of the unit tests
must be high as well.

The config part "exclude-pattern" is checked in a case insensitive
way. This means that */test/* will match an foo/tests/bar AND
Foo/Tests/bar.

Change-Id: I851b512ce53420b70af911c0dd2ebf13f8a58e48
Fixes: #37320
Releases: 0.0.4